pyproject.toml
[project]
name = "dcfriend-legacy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.13",
    "dependency-injector>=4.47.1",
    "discord-py>=2.5.2",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "pydantic>=2.0",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
pythonpath = [
    ".",
    "src",
]
asyncio_mode = "auto"
